pyproject.toml
[project]
name = "calendar-assistant-backend"
version = "0.1.0"
description = "FastAPI backend for AI-powered calendar assistant"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.8",
  "pydantic-settings>=2.4",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "httpx>=0.27",
  "python-dateutil>=2.9",
  "dateparser>=1.2",
  "PyJWT>=2.9",
  "google-auth>=2.33",
  "google-auth-oauthlib>=1.2",
  "google-api-python-client>=2.142",
  "openai>=1.43",
  "anyio>=4.4"
]

[dependency-groups]
dev = [
  "ruff>=0.6",
  "pytest>=8.3",
]

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B","SIM"]
ignore = ["E203","E501"]
